requirements_cpu.txt
affiliation @ file:///home/portege27/EurProjects/INR/code/coinpp_ad/affiliation-metrics-py
altair==4.2.2
appdirs==1.4.4
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
datatable==1.0.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
docker-pycreds==0.4.0
entrypoints==0.4
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.9.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
h5py==3.8.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython @ file:///croot/ipython_1680701871216/work
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
packaging==23.1
pandas==2.0.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
patsy==0.5.3
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
plotly==5.14.1
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.3
psutil==5.9.5
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments @ file:///croot/pygments_1683671804183/work
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.1
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.23.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels==0.14.0
streamlit==1.22.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tornado==6.3.2
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.13
validators==0.20.0
wandb==0.15.2
watchdog==3.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
zipp==3.15.0
